Update to use CMake
The <https://github.com/stripe/cpp-interview-prep> that we send to
candidates instructs them how to use CMake to set up their environment.

C++ candidates are going to need CMake anyways to build the bugsquash
question, so I suggest we be consistent with that for the integration.

Also, there are some nice things that using CMake gives (picked up by
more IDEs, easy to use to generate compile_commands.json for clangd,
etc.).