Remove WebRequest as a recommended library for c#
Microsoft has deprecated that library:
`We don't recommend that you use WebRequest or its derived classes for new development. Instead, use the System.Net.Http.HttpClient class.`

I also updated the link for HttpClient to point to the latest version of the .net framework instead of an older asp.net version. The previous link showed an example using `PostAsJsonAsync` which is not available anymore.